Refactor cleaning of existing variables of existing SOEP variables into single directory.
